float dosis

+++ cHAZEXTERNO.h
@@ -8,12 +8,12 @@ class cHAZEXTERNO :
     public cRADIOTERAPIA
 {
 public:
-    static unsigned int dosisMaxPaciente;
-    static unsigned int dosisMaxTumor;
+    static float dosisMaxPaciente;
+    static float dosisMaxTumor;
 
     cHAZEXTERNO(eTamanioTumor tamanio);
     ~cHAZEXTERNO();
-    unsigned int ajustar_dosis(eTamanioTumor tamanio);
+    float ajustar_dosis(eTamanioTumor tamanio);
     string to_string();
 };
